# release-21.2: sql: avoid string formatting in reportSessionDataChanges when not necessary